CBP-29 - Styles Refactor (#16)
* refactor: moves the styles to a global scss file

* refactor: routes and file structure
