+++ src/AppRouter.tsx
@@ -1,16 +1,16 @@
 import { Routes, Route } from "react-router-dom";
 
-import Shop from "./components/shop/shop";
-import HomePage from "./components/portfolio/HomePage";
-import DetailPage from "./components/shop/DetailPage/DetailPage";
+import HomePage from "./pages/HomePage";
+// import DetailPage from "./components/shop/DetailPage/DetailPage";
+// import Shop from "./components/shop/shop";
 
 export default function AppRouter() {
   return (
     <>
       <Routes>
         <Route path="/" element={<HomePage />} />
-        <Route path="/shop" element={<Shop />} />
-        <Route path="/shop/:itemId" element={<DetailPage />} />
+        {/* <Route path="/shop" element={<Shop />} />
+        <Route path="/shop/:itemId" element={<DetailPage />} /> */}
       </Routes>
     </>
   );
