+++ src/components/Footer.tsx
@@ -1,0 +1,9 @@
+import SocialMediaSection from "./socialMediaLinks/SocialMediaSection";
+export default function Footer() {
+  return (
+    <footer>
+      <h3> Colorblind Pixel </h3>
+      <SocialMediaSection />
+    </footer>
+  );
+}
