+++ src/components/Menu.tsx
@@ -1,4 +1,3 @@
-import "./menu.scss";
 import { Link } from "react-router-dom";
 
 // Define the type for the props
@@ -24,18 +23,18 @@ export default function Menu({ menuOpen, setMenuOpen }: MenuProps) {
           <Link to="shop">Shop</Link>
         </li>
         <div className="image-container">
-        <img
-          src="https://res.cloudinary.com/dkhpxyxnt/image/upload/c_thumb,w_200,g_face/v1726200029/zhoulogo_otm7rq.png"
-          alt="Menu Logo"
-        />
-      </div>
-      <div className="container">
-        <img
-          src="https://res.cloudinary.com/drfhleop1/image/upload/c_thumb,w_200,g_face/v1726212149/COLORBLINDTEXT_bcjb5u.png"
-          alt="Brand Name Image"
-          className="moving-image"
-        />
-      </div>
+          <img
+            src="https://res.cloudinary.com/dkhpxyxnt/image/upload/c_thumb,w_200,g_face/v1726200029/zhoulogo_otm7rq.png"
+            alt="Menu Logo"
+          />
+        </div>
+        <div className="container">
+          <img
+            src="https://res.cloudinary.com/drfhleop1/image/upload/c_thumb,w_200,g_face/v1726212149/COLORBLINDTEXT_bcjb5u.png"
+            alt="Brand Name Image"
+            className="moving-image"
+          />
+        </div>
         {/* 
         <li onClick={() => setMenuOpen(false)}>
           <Link to="shop">Shop</Link>
