--- src/components/PortfolioCategories/PortfolioCategories.scss
@@ -1,20 +1,0 @@
-@import '../../global.scss';
-
-.portfolio-categories {
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-  margin: 1rem 0;
-
-  .button {
-  background-color: lighten($mainColor, 35%);
-    &.active {
-      background-color: $mainColor;
-      color: #fff;
-    }
-
-    &:hover {
-      background-color: lighten($mainColor, 10%);
-    }
-  }
-}
