+++ src/components/Topbar.tsx
@@ -1,4 +1,3 @@
-import "./topbar.scss";
 import { Link } from "react-router-dom";
 
 interface TopbarProps {
