--- src/components/contact/contact.scss
@@ -1,127 +1,0 @@
-@import "../../global.scss";
-
-.contact {
-  background-color: white;
-  display: flex;
-
-  @include mobile {
-    flex-direction: column;
-  }
-
-  .left {
-    @include mobile {
-      display: none;
-    }
-    flex: 1;
-    overflow: hidden;
-
-    img {
-      height: 100%;
-    }
-  }
-  .right {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    h2 {
-      font-size: 30px;
-      margin-bottom: 20px;
-    }
-
-    form {
-      width: 70%;
-      height: 50%;
-      flex-direction: column;
-      align-items: center;
-      border-radius: 10px;
-      box-shadow: 0px 0px 15px -8px black;
-      display: flex;
-      flex-direction: column;
-      padding: 20px;
-
-      * {
-        overflow: hidden;
-      }
-
-      input {
-        width: 300px;
-        height: 30px;
-        font-size: 14px;
-        border: none;
-        border-radius: 10px;
-        background-color: rgb(238, 229, 229);
-        margin: 10px;
-        padding: 5px;
-        @include mobile {
-          width: 200px;
-          height: 20px;
-        }
-      }
-
-      textarea {
-        border: none;
-        width: 300px;
-        height: 200px;
-        font-size: 14px;
-        border-radius: 10px;
-        background-color: rgb(238, 229, 229);
-        margin: 10px;
-        padding: 5px;
-        @include mobile {
-          width: 200px;
-          height: 100px;
-        }
-      }
-
-      button {
-        overflow: hidden;
-        width: 150px;
-        height: 30px;
-        color: white;
-        background-color: rgb(0, 0, 0);
-        border: none;
-        border-radius: 10px;
-        font-weight: 500;
-        cursor: pointer;
-
-        &:focus {
-          outline: none;
-        }
-      }
-
-      button:hover {
-        background-color: crimson;
-      }
-
-      span {
-        padding-top: 20px;
-        color: green;
-      }
-
-      h3 {
-        padding-top: 20px;
-        align-items: center;
-        justify-content: center;
-      }
-
-      .socialMediaLogos {
-        margin-top: 10px;
-        flex-direction: row;
-        display: flex;
-        margin: 0;
-        padding: 0;
-        .fa {
-          padding: 10px;
-          font-size: 30px;
-          width: 30px;
-          text-align: center;
-          text-decoration: none;
-          border-radius: 50%;
-        }
-      }
-    }
-  }
-}
