--- src/components/footer/Footer.tsx
@@ -1,11 +1,0 @@
-import "./footer.scss"
-import SocialMediaSection from "../socialMediaLinks/SocialMediaSection";
-export default function Footer() {
-  return (
-    <footer>
-    
-    <h3> Colorblind Pixel </h3>
-    <SocialMediaSection />
-    </footer>
-  )
-}
