--- src/components/footer/footer.scss
@@ -1,21 +1,0 @@
-@import "../../global.scss";
-
-footer {
-  background-color: black;
-  flex: none;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 40px;
-  max-height: 150px;
-  h3 {
-    color: white;
-  }
-
-  .socialMediaLogos {
-    flex-direction: row;
-    display: flex;
-    justify-content: left;
-  }
-}
