--- src/components/intro/intro.scss
@@ -1,150 +1,0 @@
-@import "../../global.scss";
-
-.intro {
-  width: 100vw;
-  height: 100vh;
-  background-color: white;
-  display: flex;
-  overflow: hidden;
-
-  @include mobile {
-    flex-direction: column;
-  }
-
-  .left {
-    flex: 0.5;
-    display: flex;
-    overflow: hidden;
-    align-items: center;
-    justify-content: center;
-
-    .imgContainer {
-     
-      width: 100vh;
-      height: 100vh;
-      background-color: black;
-      //border-radius: 50%;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      @include mobile {
-        margin-top: 0;
-        width: 100%;
-        height: 100%;
-        border-radius: 0%;
-      }
-      @include ultrawide {
-        height: 100%;
-        width: 100%;
-      }
-
-      img {
-        height: 100%;
-        width: 100%;
-        object-fit: cover;
-        @include mobile {
-          height: 100%;
-          width: 100%;
-        }
-      }
-    }
-  }
-  .right {
-    flex: 0.5;
-    display: flex;
-    position: relative;
-    justify-content: center;
-    align-items: center;
-    .wrapper {
-      width: 100%;
-      height: 100%;
-      padding-left: 100px;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      position: absolute;
-
-      @include mobile {
-        padding-left: 0;
-        align-items: center;
-      }
-
-      h1 {
-        font-size: 60px;
-        margin: 10px 0;
-
-        @include mobile {
-          font-size: 40px;
-        }
-
-        @include ultrawide {
-          font-size: 150px;
-        }
-  
-      }
-      h2 {
-        font-size: 35px;
-        @include ultrawide {
-          font-size: 60px;
-        }
-      }
-      h3 {
-        font-size: 30px;
-
-        @include mobile {
-          font-size: 20px;
-        }
-
-        @include ultrawide {
-          font-size: 60px;
-        }
-
-        span {
-          font-size: inherit;
-          color: crimson;
-        }
-
-        .ityped-cursor {
-          animation: blink 1s infinite;
-        }
-
-        @keyframes blink {
-          50% {
-            opacity: 1;
-          }
-          100% {
-            opacity: 0;
-          }
-        }
-      }
-      .socialMediaLogos{
-        padding-top: 10px;
-        flex-direction: row;
-        display: flex;
-        justify-content: left;
-      }
-    }
-    .arrow{
-      a {
-        @include mobile {
-          justify-content: center;
-          opacity: 0;
-        }
-        position: absolute;
-        bottom: 10px;
-  
-        img {
-          width: 30px;
-          animation: arrowBlink 2s infinite;
-        }
-  
-        @keyframes arrowBlink {
-          100% {
-            opacity: 0;
-          }
-        }
-      }
-    }
-  }
-}
