--- src/components/menu/menu.scss
@@ -1,72 +1,0 @@
-@import "../../global.scss";
-
-.menu{
-    width: 400px;
-    height: 100vh;
-    background-color: $mainColor;
-    position: fixed;
-    top: 0;
-    right: -400px;
-    z-index: 6;
-    display: flex;
-    flex-direction: column;
-    align-items:  center;
-    transition: all 1s ease;
-    &.active{
-        right: 0;
-    }
-
-    ul{
-        margin-top: 8rem;
-        padding: 0;
-        list-style: none;
-        font-size: 30px;
-        font-weight: 200;
-        color: white;
-        width: 60%;
-        
-        li{
-            margin-bottom: 25px;
-            a{
-                font-size: 30px;
-                color: inherit;
-                text-decoration: none;
-            }
-
-            &:hover {
-                color: yellowgreen;
-            }
-        }
-    }
-}
-
-// LOGO IMAGE
-.menu img {
-    display: block;
-    margin: 5rem auto;
-  }
-  .menu .image-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  // ANIMATION
-  .moving-image {
-    position: relative;
-    width: 200px; /* Adjust as needed */
-    animation: moveRightToLeft 5s linear infinite;
-  }
-  @keyframes moveRightToLeft {
-    0% {
-      left: 100%;  /* Start from the right */
-    }
-    100% {
-      left: -100%; /* Move to the left */
-    }
-  }
-  .container {
-    overflow: hidden;
-    width: 100%;
-  }
-
