--- src/components/portfolio/homePage.scss
@@ -1,39 +1,0 @@
-@import "../../global.scss";
-.app {
-  padding: 0;
-  margin: 0;
-  height: 100vh;
-  flex-grow: 1;
-
-  .sections {
-    padding: 0;
-    margin: 0;
-    width: 100vw;
-    min-height: calc(100vh - 70px);
-    top: 70px;
-    scroll-snap-type: y mandatory;
-    scroll-behavior: smooth;
-    scrollbar-width: none; //for firefox
-    &::-webkit-scrollbar {
-      display: none;
-    }
-
-    > * {
-      width: 100vw;
-      min-height: calc(100vh - 70px);
-    }
-
-    .portfolio {
-      height: auto;
-      padding-bottom: 30px;
-    }
-
-    @include mobile {
-      top: 50px;
-      > * {
-        width: 100vw;
-        min-height: calc(100vh - 70px);
-      }
-    }
-  }
-}
