--- src/components/portfolio/portfolio.scss
@@ -1,7 +1,0 @@
-@import "../../global.scss";
-
-.portfolio-load-more {
-  display: flex;
-  justify-content: center;
-  margin: 1rem 0;
-}
