--- src/components/portfolioTabs/portfolioTabs.scss
@@ -1,21 +1,0 @@
-@import "../../global.scss";
-
-.portfolioList {
-  font-size: 14px;
-  margin-right: 50px;
-  padding: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  overflow: hidden;
-
-
-  @include mobile{
-    margin: 10px;
-  }
-
-  &.active {
-    border-bottom: 1px solid black;
-  }
-}
