--- src/components/socialMediaLinks/icon.scss
@@ -1,28 +1,0 @@
-.icon{
-  padding: 10px;
-  .fa {
-    padding: 20px;
-    font-size: 30px;
-    width: 30px;
-    text-align: center;
-    text-decoration: none;
-    border-radius: 50%;
-  }
-  
-  .fa-facebook {
-      background: #000000;
-      color: white;
-    }
-
-  .fa-instagram {
-    background: #000000;
-    color: white;
-  }
-
-  .fa-twitter {
-    background: #000000;
-    color: white;
-  }
-
-  
-}
