--- src/components/topbar/topbar.scss
@@ -1,158 +1,0 @@
-@import "../../global.scss";
-.topbar {
-  width: 100%;
-  padding: 0px;
-  margin: 0px;
-  height: 70px;
-  background-color: white;
-  color: $mainColor;
-  position: fixed;
-  z-index: 1000;
-  transition: all 1s ease;
-  overflow: hidden;
-
-  @include mobile {
-    height: 50px;
-  }
-  .wrapper {
-    padding: 10px 30px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    .left {
-      display: flex;
-      align-items: center;
-      .logo {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        font-size: 40px;
-        font-weight: 700;
-        text-decoration: none;
-        color: inherit;
-        margin-right: 40px;
-
-        @include mobile {
-          font-size: 20px;
-          margin-right: 0px;
-        }
-
-        img {
-          height: 1em;
-          margin-right: 5px;
-
-          @include mobile {
-            height: 1.5em;
-          }
-        }
-      }
-
-      .itemContainer {
-        display: flex;
-        align-items: center;
-        margin-left: 30px;
-
-        @include mobile {
-          display: none;
-        }
-        .icon {
-          font-size: 18px;
-          margin-right: 5px;
-        }
-
-        span {
-          font-size: 25px;
-          font-weight: 400;
-        }
-      }
-
-      span:hover {
-         border-bottom: 1px solid black;
-      }
-    }
-
-    button {
-      overflow: hidden;
-      width: 170px;
-      height: 40px;
-      color: white;
-      padding: 20px;
-      background-color: rgb(0, 0, 0);
-      border: none;
-      border-radius: 10px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      a {
-        font-weight: 500;
-        color: white;
-        text-decoration: none;
-        font-size: 20px;
-      }
-
-      cursor: pointer;
-
-      &:focus {
-        outline: none;
-      }
-
-      @include mobile {
-        display: none;
-      }
-    }
-
-    button:hover {
-      background-color: crimson;
-    }
-
-    .right {
-      display: none;
-      @include mobile {
-        display: flex;
-      }
-      .hamburger {
-        width: 32px;
-        height: 25px;
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-        cursor: pointer;
-
-        span {
-          width: 100%;
-          height: 3px;
-          background-color: $mainColor;
-          transform-origin: left;
-          transition: all 2s ease;
-        }
-      }
-    }
-  }
-
-  &.active {
-    background-color: $mainColor;
-    color: white;
-
-    .hamburger {
-      scrollbar-width: none; //for firefox
-      &::-webkit-scrollbar {
-        display: none;
-      }
-      span {
-        &:first-child {
-          background-color: white;
-          transform: rotate(45deg);
-        }
-        &:nth-child(2) {
-          opacity: 0;
-        }
-        &:last-child {
-          background-color: white;
-          transform: rotate(-45deg);
-        }
-      }
-    }
-  }
-}
