+++ src/pages/DetailPage/DetailPage.tsx
