+++ src/pages/DetailPage/detailPage.scss
