+++ src/pages/shop/ItemCard/itemCard.scss
