+++ src/pages/shop/shop.tsx
@@ -1,5 +1,5 @@
 import "./shop.scss";
-import Footer from "../footer/Footer";
+import Footer from "../../components/Footer";
 import ItemCard from "./ItemCard/ItemCard";
 
 import { useState, useEffect } from "react";
@@ -25,7 +25,8 @@ export default function Shop(): JSX.Element {
           Name: "Grumpy Cat Poster",
           Description: "Everyone's favorite cat who loves to hate",
           Price: 15,
-          ImageURL://link works correctly, this one is just not reachable. Using this as placeholder
+          //link works correctly, this one is just not reachable. Using this as placeholder
+          ImageURL:
             "https://m.media-amazon.com/images/I/71-3ZzX-QLL._AC_UF894,1000_QL80_.jpg",
           Amount: 0,
         },
